Refactor battle modules into separate components
