+++ pokemon/data/learnsets/__init__.py
@@ -1,0 +1,1 @@
+"""Learnset data package."""
